ramips: add support for Z-router ZR-2662
WIP

Without lzma-loader the kernel does not boot:

 Uncompressing Kernel Image ... lzma compressed: uncompress error 1
 Must RESET board to recover

Changing LINUX_LOADADDR to 0x81001000 as used by OEM solves the lzma
error but then the kernel doesn't boot - no output at all on the console
after this line:

 Loading Device Tree to 9be77000, end 9be7c4b7 ... OK

Eventually came accross this forum post:
https://forum.openwrt.org/t/compile-ramips-with-data-address-at-0x81001000-instead-of-0x80001000/97379

Which seems solve the problem. The question is: is there a better way?
Can't find any documentation about lzma-loader, what it does exactly,
and why it is needed, so for now this allows to boot OpenWrt from flash.